+++ Guitar++/CPlayer.h
@@ -332,7 +332,7 @@ public:
 	bool lastFretsPressed[5];
 	bool fretsPressed[5];
 	double fretsPressedTime[5];
-	int64_t notesSlide[5];
+	size_t notesSlide[5];
 
 	size_t strklinent;
 	double strklinenttime;
